requirements.txt
telnyx==2.1.3
pyyaml==6.0.1
pytz==2024.1
python-dotenv==1.0.1
